+++ src/bin/main.rs
@@ -1,23 +1,10 @@
 //! A re-implementation of the `electron-evil-feature-patcher` CLI tool that works nearly identically.
 
-use electron_hardener::{
-    patcher::{DevToolsMessage, ElectronOption, NodeJsCommandLineFlag},
-    ElectronApp, Fuse,
-};
+use electron_hardener::{patcher::ElectronOption, ElectronApp, Fuse};
 use std::{env, fs};
 
-const FUSES: &[Fuse] = &[Fuse::RunAsNode];
-
-const NODEJS_FLAGS: &[NodeJsCommandLineFlag] = &[
-    NodeJsCommandLineFlag::Inspect,
-    NodeJsCommandLineFlag::InspectBrk,
-    NodeJsCommandLineFlag::InspectPort,
-    NodeJsCommandLineFlag::Debug,
-    NodeJsCommandLineFlag::DebugBrk,
-    NodeJsCommandLineFlag::DebugPort,
-    NodeJsCommandLineFlag::InspectBrkNode,
-    NodeJsCommandLineFlag::InspectPublishUid,
-];
+const FUSES_TO_DISABLE: &[Fuse] = &[Fuse::RunAsNode, Fuse::NodeOptions, Fuse::NodeCliInspect];
+const FUSES_TO_ENABLE: &[Fuse] = &[Fuse::OnlyLoadAppFromAsar];
 
 const ELECTRON_FLAGS: &[ElectronOption] = &[
     ElectronOption::JsFlags,
@@ -25,9 +12,6 @@ const ELECTRON_FLAGS: &[ElectronOption] = &[
     ElectronOption::RemoteDebuggingPort,
     ElectronOption::WaitForDebuggerChildren,
 ];
-
-const DEVTOOLS_MESSAGES: &[DevToolsMessage] =
-    &[DevToolsMessage::Listening, DevToolsMessage::ListeningWs];
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let application_path = env::args()
@@ -38,20 +22,16 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let mut app = ElectronApp::from_bytes(&mut application_bytes)?;
 
-    for fuse in FUSES.iter().copied() {
+    for fuse in FUSES_TO_DISABLE.iter().copied() {
         app.set_fuse_status(fuse, false)?;
     }
 
-    for flag in NODEJS_FLAGS.iter().copied() {
-        app.patch_option(flag)?;
+    for fuse in FUSES_TO_ENABLE.iter().copied() {
+        app.set_fuse_status(fuse, true)?;
     }
 
     for flag in ELECTRON_FLAGS.iter().copied() {
         app.patch_option(flag)?;
-    }
-
-    for msg in DEVTOOLS_MESSAGES.iter().copied() {
-        app.patch_option(msg)?;
     }
 
     fs::write(application_path, application_bytes)?;